added working correction test using optix ray casting correspondences and reduction on GPU

+++ src/rmagine_core/include/rmagine/types/PointCloud.hpp
@@ -51,7 +51,7 @@ template<typename MemT>
 struct PointCloud_
 {
     Memory<Vector, MemT>        points;
-    Memory<unsigned int, MemT>  mask;
+    Memory<uint8_t, MemT>       mask;
     Memory<Vector, MemT>        normals;
     Memory<unsigned int, MemT>  ids;
 };
@@ -62,9 +62,9 @@ template<typename MemT>
 struct PointCloudView_
 {
     MemoryView<Vector, MemT>        points; // required
-    MemoryView<unsigned int, MemT>  mask    = MemoryView<unsigned int, MemT>::Empty();
+    MemoryView<uint8_t, MemT>       mask    = MemoryView<uint8_t, MemT>::Empty();
     MemoryView<Vector, MemT>        normals = MemoryView<Vector, MemT>::Empty();
-    MemoryView<unsigned int, MemT>  ids     = MemoryView<unsigned int, MemT>::Empty();
+    MemoryView<unsigned int, MemT>  ids     = MemoryView<uint32_t, MemT>::Empty();
 };
 
 // default: RAM
